requirements.txt
rembg
Pillow
numpy
onnxruntime